scsi: lpfc: Refresh ndlp when a new PRLI is received in the
From ecf041fe9895c3f5c0119fb427fa634a275c9382 Mon Sep 17 00:00:00 2001
 PRLI issue state
Git-commit: ecf041fe9895c3f5c0119fb427fa634a275c9382
Patch-mainline: v5.12-rc2
References: bsn#61

Testing with target ports coming and going, the driver eventually reached a
state where it no longer discovered the target. When the driver has issued
a PRLI and receives a PRLI from the target, it is not properly updating the
node's initiator/target role flags. Thus, when a subsequent RSCN is
received for a target loss, the driver mis-identifies the target as an
initiator and does not initiate LUN scanning.

Fix by always refreshing the ndlp with the latest PRLI state information
whenever a PRLI is processed.  Also clear the ndlp flags when processing a
PLOGI so that there is no carry over through a re-login.

Link: https://lore.kernel.org/r/20210104180240.46824-4-jsmart2021@gmail.com
Co-developed-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: Dick Kennedy <dick.kennedy@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.12-rc2-scsi-lpfc-Refresh-ndlp-when-a-new-PRLI-is-received-i.patch