ASoC: rt5682: fix getting the wrong device id when the
Git-commit: 696cfb25679452721215c193ba6f72eebfe7a842
 suspend_stress_test
References: bsn#19
Patch-mainline: v5.10.50

[ Upstream commit 867f8d18df4f5ccd6c2daf4441a6adeca0b9725b ]

This patch will be the workaround to fix getting the wrong device ID on the rare chance.
It seems like something unstable when the system resumes. e.g. the bus clock
This patch tries to read the device ID to check several times.
After the test, the driver will get the correct device ID the second time.

Signed-off-by: Shuming Fan <shumingf@realtek.com>
Link: https://lore.kernel.org/r/20210111092740.9128-1-shumingf@realtek.com
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.50-0507-ASoC-rt5682-fix-getting-the-wrong-device-id-when-the.patch